+++ package.json
@@ -17,5 +17,8 @@
   ],
   "devDependencies": {
     "concurrently": "^8.2.1"
+  },
+  "dependencies": {
+    "react-router-dom": "^6.20.1"
   }
 }
